[1.11] Fix setting etcd-server when no mTLS is set between etcd and kube-apiserver
An error is introduced in commit 84598f73b9f6881958370061dbfda49d4d7c396b when I cherry pick https://github.com/kubernetes/kubernetes/pull/74690, when mTLS is not enabled, etcd-server was not set, this patch is to fix the issue. This PR is temporarily rolled back in OSS for performance issue investigation, so I cannot fix it in OSS now. I am fixing this in GKE for now, I will add a buganizer to track and make sure when this is fixed in open source, this fix will not carry on.

Change-Id: I8f50e333a3a394e4cad8cdb96a9a31fe9112cb61