Support workload identity on GKE Windows node.
It first test if workload identity is enabled. If true, it downloads
gke-metadata-server.exe and starts it as native Windows service on
cbr0:988. Its metrics are exported on MGMT_IP:989 as gke-metrics-agent
scrape metrics on MGMT_IP.

Proxy policy will be applied to all pods on this node to redirect
169.254.169.254:80 to cbr0:988.
Access control list policy will be applied to cbr0 so that cbr0:988 can't be
accessed from outside the node.

Add a new kube_env ENABLE_WINDOWS_WORKLOAD_IDENTITY to control this
feature. After this PR is merged and cherry-picked, we can start to
allow creating Windows node with workload identity enabled and set it.

Ref: https://cloud.google.com/kubernetes-engine/docs/how-to/workload-identity
Bug: 154748130
Guitar tests bug: 159131141

Change-Id: I8a638336b5546b01acdd4e52d98528d919fa8a3d
(cherry picked from commit 4655908715041ce5092fc001360a023450f807b0)